pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "..."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "alembic>=1.17.1",
    "fastapi>=0.121.1",
    "passlib[argon2,bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "pydantic[email]",
    "uvicorn[standard]>=0.38.0",
    "authx>=1.4.3",
    "jinja2>=3.1.6",
]
